Getting the Token via Client Credentials
Put the champagne away: we're not done yet. Typically, access tokens don't last forever. COOP tokens last for 24 hours, which means that tomorrow, our script will break.

Letting the website do the client-credentials work for us was nice for testing, but we need to do it ourselves inside the script. Every OAuth server has an API endpoint used to request access tokens. If we look at the COOP API Authentication docs, we can see the URL and the POST parameters it needs:
http://coop.apps.knpuniversity.com/token

Parameters:
    client_id
    client_secret
    grant_type
Let's update our script to first make this API request. Fill in the client_id, client_secret and grant_type POST parameters.

With any luck, when you run it, you should see a JSON response with an access token and a few other details:

{
  'access_token': 'fa3b4e29d8df9900816547b8e53f87034893d84c',
  'expires_in': 86400,
  'token_type': 'Bearer',
  'scope': 'chickens-feed'
}