Let's set that URL to be /coop/oauth/handle, which is just another page that's printing a message. The code for this is right inside the same file, a little further down. Instead of hardcoding the URL, I'll use the URL generator that's part of Silex. However you make your URL, just make sure it's absolute. Ok, we've built our authorize URL to COOP, let's redirect the user to it.
That redirect() function is special to my app, so your code may differ. As long as you somehow redirect the user, you're good.

 Tip
Since we're using Silex, the redirect() function is actually a shortcut I created to create a new RedirectResponse object.

Authorizing on COOP

Let's try it! Go back to the homepage and click the 'Authorize' link. This takes us to our code, which then redirects us to COOP. We're already logged in, so it gets straight to asking us to authorize the app. Notice that the scopes that we included in the URL are clearly communicated. Let's authorize the app. Later, we'll see what happens if you don't.

When we click the authorization button, we're sent back to the redirect_uri on TopCluck! Nothing has really happened yet. COOP didn't set any cookies or anything else. But the URL does include a code query parameter.