What if we want to make other API requests on behalf of Brent later? Where should we store the access token?
Saving the Access Token Somewhere

Some access tokens last an hour or two, and are well suited for storing in the session. Others are long-term tokens, for example facebook provides a 60-day token, and these make more sense to store in a database. Either way, storing the token will free us from having to ask the user to authorize again.

In our app, we're going to store it in the database.
This code is specific to my app, but the end result is that I've updated the coopAccessToken column on the user table for the currently-authenticated user. I'm also saving the coopUserId, which we'll need since most API calls have the user's ID in the URI.