Recording the Expires Time
We can also store the time when the token will expire. I'll create a DateTime object that represents the expiration time. We can check this later before trying to make API requests. If the token is expired, we'll need to send the user through the authorization process again.

Again, the code here is special to my app, but the end result is just to update a column in the database for the current user. When we try it, it runs and hits our die statement. But if you go to the homepage, the user drop-down shows us that the COOP user id was saved! Eggcellent...