When Fetching the Access Token Fails
There's one other spot where things can fail: when requesting out to /token. What if the response doesn't have an access_token field? Under normal circumstances, this really shouldn't happen, but let's render a different error template in case it does. Don't worry about the variables I'm passing into the template, I'm just trying to pass enough information so that we can see what the problem was.

Try the whole cycle again, but approve the app this time. It works the first time of course. But if you refresh, you'll see this error in action. The code parameter exists, but it's expired. So, the request to /token fails.

Redirecting after Success

Until now, we've had an ugly die statement at the bottom of the code that handles the OAuth redirect. What you'll actually want to do here is redirect to some other page. Our work is done for now, so we want to help the user to continue on our site.

In our application, this code simply redirects us to the homepage. And just like that, we're done! This is the authorization grant type, which has 2 distinct steps to it:

 1.First, redirect the user to the OAuth server using its /authorize endpoint, your application's client_id, a redirect_uri and the scopes you want permission for. The URL and how the parameters look may be different on other OAuth servers, but the idea will be the same.

 2.After authorizing our app, the OAuth server redirects back to a URL on our site with a code query parameter. We can use this, along with our client_id and client_secret to make an API request to the /token endpoint. Now, we have an access token.

Let's finally use it to count some eggs!