Couting Eggs
On the homepage, we still have the 'Authorize' button. But now that we have an access token for the user, we really don't need this anymore. The template that displays this page is at views/dashboard.twig, and I'm already passing a user variable here, which is the currently-authenticated user object. Let's hide the 'Authorize' link if the user has a coopUserId stored in the database.
If we do have a coopUserId, let's add a link the user can click that will count their daily eggs. Don't worry if you're not familiar with the code here, we're just generating a URL to a new page that I've already setup.

When we refresh, we see the new link. Clicking it gives us another todo message.