Open up src/OAuth2Demo/Client/Controllers/CountEggs.php, which is the code behind this new page.
Making the eggs-count API Request

Start by copying the /api/me code from CoopOAuthController, and changing the method from get() to post(), since the eggs-count endpoint requires POST.

The endpoint we want to hit now is /api/USER_ID/eggs-count. Fortunately, we've already saved the COOP user id and access token for the currently logged-in user to the database. Get that data by using our app's {dollar_sign}this->getLoggedInUser() method and update the URL.

I'll add in some debug code so we can see if this is working.

When we refresh, you should see a nice JSON response. Yea, we're counting eggs! That'll show Farmer Scott!