All the Things that can Go Wrong
The 'Count Eggs' page we created works great, but we're not handling any of the things that might go wrong. First, we're hiding its link, but what if a user somehow ends up on the page without a coopUserId or coopAccessToken? Let's code for this case.

I'm throwing an exception message, but we could also handle this differently, like by redirecting the user to the 'Authorize' page to start the OAuth flow.

Another thing we can check for is whether or not the token has expired. This is possible because we stored the expiration data in the database. I've created an easy helper method to check for this. If this happens, let's redirect the user to re-authorize, just like if they had clicked the 'Authorize' link.

Finally, what if the API request itself fails? A simple way to handle this might look like this.

Of course, you may want to do something more sophisticated. The response could also have some error information on it, which you can play around with. For OAuth, this is important because the call may have failed because the access_token expired. What, I thought we just checked for that? Well, in the real world, there's no guarantee that the token won't expire before its scheduled time. Plus, the user may have decided to revoke your token -- what a bully. Be aware, and handle accordingly. Once again, the OAuth Server should provide information on the error in the error and error_description query string parameters.

You're now dangerous, so lets move on to let our farmers actualy log into FCL via COOP.