Handling Existing Users
There was one big hole in our logic. If I had logout and go through the process again, it blowed up! That time, it tried to create a second new user for Brent instead of using the one from before. Let's fix that. For organization, I'm going to create a new private function called findOrCreateUser() in this same class. If we can find a user with this COOP User ID, then we can just log the user into that account. If not, we'll keep creating a new one.

Try the process again. No error this time - we find the existing user and use it instead of creating a new one.