Now that Brent has TopCluck up and running he can finally challenge Farmer Scott to an all out egg counting brawl. Brent knows that if they are both tracking egg collections on TopCluck Farmer Scott will be proven wrong and everyone will see how awesome his hens really are.
But what fun is winning if no one else gets to see? So Brent hatches another idea: having users share their chicken-laying progress on Facebook.

Fortunately, Facebook uses OAuth 2.0 for their API, so we're already dangerous. And like a lot of sites, they even have a PHP library to help us work with it. Installing it via Composer is easy. In fact, I already added it to our composer.json, so the library is downloaded and ready to go:

{
  'require': {
    ...
    'facebook/graph-sdk': '^5.5'
  }
}
The library has a simple example, but it's easier to see it integrated in a real application.

The FacebookOAuthController

We're going to integrate with Facebook using the same authorization code grant type we just used with COOP. So it shouldn't be any surprise that we need the same 2 pages as before: 1 that redirects to Facebook and 1 that handles things after Facebook redirects back to us.

In fact, if you open up FacebookOAuthController.php, you'll see that I've started us with a setup that looks exactly like we had with COOP.

Starting the Redirect

Let's start by adding a link on the homepage to 'Connect with Facebook'

When we click this, we hit the code in the first function. Just like before, our job is to redirect to the authorize URL on Facebook. If we dig a little(https://developers.facebook.com/docs/facebook-login/manually-build-a-login-flow/) bit on Google, we can see this is /dialog/oauth. We could start building this by hand, but the Facebook SDK can help us out.