Redirecting the User Now, to get the authorize URL, we can use the getLoginUrl()-https://developers.facebook.com/docs/reference/php/facebook-getLoginUrl/- function on the SDK. Remember that this URL always has 3 important things on it: the client ID, the redirect URI back to our site and the list of scopes we need. The object already has our client ID, so lets pass the redirect URI and scopes here. For Facebook, these are called redirect_uri and scope.
To know which scopes you need, you have to check with the API you're using. If we google about Facebook API scopes, we find a page -https://developers.facebook.com/docs/reference/login/ - that explains all of them. We'll ultimately want to be able to get basic user information and post to a user's timeline. These are email and publish_actions.

Finally, let's redirect the user to this URL. The flow should feel completely familiar by now.

Registering the Redirect URI

When we try it out, we do go to Facebook's /dialog/oauth with the client_id, redirect_uri and scope parameters. But we get an error:

Given URL is not allowed by the Application configuration. One or more
of the given URLs is not allowed by the App's settings. It must match
the Website URL or Canvas URL, or the domain must be a subdomain of one
of the App's domains.
It's complaining about the redirect URL we're sending. For added security, OAuth servers allow, and sometimes require you to configure your redirect URL in your application. Go back to our application and click Settings and then 'Add Platform'. Choose 'Website' and then fill in the URL of your site.

 Tip
Facebook likes to change their interface, so this may look different someday soon! But one way or another, you're looking for a way to register your redirect URL.

And just like that, when we try it again, it works. Facebook made us do that so that no other sites can try to use our app id and have Facebook redirect back to some other domain. COOP's application settings also have this ability, but it wasn't required, so we skipped it. But, it's always better to fill this in.

At the authorize URL, Facebook describes the scopes that we're asking for, including the ability to post. One nice thing about Facebook is that we can choose to grant this scope, but make any posts show only to us. That's a great way to test things.