Getting the Access Token When we finish, we're redirected back to our second page, which still has the original todo message. But we have a code query parameter, and we know that it can be exchanged for an access token.
Start by creating a private function that creates the Facebook object, and use it in both functions.

OAuth tells us that our next step is to make an API request to the token endpoint to exchange our authorization code for an access token. That's absolutely right, and it can be done with the help of the SDK.