When something does go wrong, Facebook will redirect back to us with information about what went wrong on the standard error and error_description query parameters. Because they're following this OAuth standard, we can easily find error details and even decide what to do next. For example, if the error is set to access_denied, then it means the user denied our authorization request. In our app, I'm just passing all of the query parameters into a template that will display them.
To try this, we first need to go to Facebook and remove the app from our account. Unlike COOP, most OAuth servers remember if you authorized an app and don't ask you again.

On TopCluck, click 'Connect with Facebook' again but 'Cancel' the authorization request. After the redirect, we see the error, error_description and error_reason query parameters. But instead of seeing the error template, our valid userId is printed out as if it were successful. What just happened?

Our OAuth flow did fail. But even still, the Facebook object looks and finds a valid access token that it stored in the session from the last, successful authorization. That's nice, but it's unexpected. Just remember that getUser() tries many things: like exchanging the authorization code for an access token or simply finding an access token that it already stored in the session.

To see the error page, clear out your session cookie to reset everything. Log back in, then connect with Facebook but deny the request again. Oh Cluck! Error page! Without any session data to fall back on, the Facebook object doesn't have an access token and so can't make an API request to get the user id.