Saving the Facebook User ID In CoopOAuthController, once we have the access token, our next step was to store some details in the database for the user, like the COOP user id, access token and expiration date.
For Facebook, I want to do something similar, but let's only store the Facebook user id. We can do this without any more work because the getUser() function gives us that id.

And of course, let's redirect back to the homepage after finishing. Try the whole cycle out - this time approving our application's authorization request. We now know that a lot is happening behind the scenes.

First, the Facebook object exchanges the authorization code for an access token and saves it in the session. This all happens when we call getUser(). Next, we save the Facebook user ID into the database and redirect to the homepage. Clicking the 'User Info' box shows us the Facebook ID.