Store the Access Token in the Database? So why aren't we storing the access token or expiration? Actually, this is up to you. The Facebook object is automatically storing the access token in the session. So, everything is easy right now.
But on the user's next session, the access token will be gone and we'll need to re-ask the user to authorize. If you want to avoid this, you could store the Facebook access token in the database. In a second, I'll show you how you'd use that access token. Of course, these tokens don't last forever, so eventually you'll need to re-authorize them or use a Using Refresh Tokens(https://knpuniversity.com/screencast/oauth/refresh-token#refresh-tokens), the topic of an upcoming chapter!