Using the Facebook API
To post to someone's timeline, we'll use Facebook's API. Like with any API that uses OAuth, we just need to know the URL, the HTTP method, any data we need to send, and how the access token should be attached to the request.

With some quick googling(https://developers.facebook.com/docs/reference/api/publishing/), we see that we need to make a POST request to /[USER_ID]/feed and send message and access_token POST data.

We could absolutely do this manually, using the nice Guzzle library from before. But since we're using the Facebook SDK, it's even easier.

Use the createFacebook method from before to get our Facebook object. Create also a FacebookApp object and a FacebookRequest object. We must attach the access token to our FacebookRequest. We may obtain this from our FacebookApp object using getAccessToken method.
Our FacebookRequest object takes 4 arguments, our FacebookApp object, the access token, the method and any parameter you send to the method.

Remember that one of the reasons this works is that our authorization URL included the scope publish_actions. Had we not done that, this request would fail.

Let's make the message more realistic by putting in my egg count and finish the flow by redirecting back to the homepage.

Refresh to try it all again. Check Facebook to see that we're bragging about our egg-laying hens' progress!