Re-trying an API Request
Depending on the error, you might also want to re-try the request. Let's refactor the API call into a new private method called makeApiRequest().
This method does the exact same thing as before. The if statement checks to see if makeApiRequest() needs us to redirect the user back to the authorize URL.

But if we add a new {dollar_sign}retry argument, we could run the request 1 more time if it fails.

Of course, this is really only interesting if we expect Facebook to have a decent number of temporary failures. But the big idea is that you should do your best to figure out why a failure has happened and re-try if it makes sense.

Tip
If you're using the Guzzle(http://guzzle.readthedocs.org/) library to make API requests (which the Facebook class does not use), it has built-in support for re-trying a request if it fails. See Guzzle Retry Subscriber(https://github.com/guzzle/retry-subscriber) (for Guzzle version 4).

This is especially useful in the world of OAuth. We didn't store the Facebook access token in the database. But if we had, we could use it right now and re-try the request again.
So if the access token were missing from the session and the one in the database hasn't expired, this will make everything work perfectly smooth.